Condition model added

+++ app/src/main/java/com/example/hw1_mas/models/Day.java
@@ -6,14 +6,14 @@ public class Day {
 //    private Date date = new Date();
     private String date;
     @SerializedName("avgtemp_c")
-    private int averageTemp ;
+    private double averageTemp ;
     private Condition condition;
 
     public String getDate() {
         return date;
     }
 
-    public int getAverageTemp() {
+    public double getAverageTemp() {
         return averageTemp;
     }
